Initial refactorings
