+++ lib/darwin/chrome-canary.js
