+++ lib/darwin/index.js
@@ -2,6 +2,6 @@ exports.safari = require( './safari' );
 exports.firefox = require( './firefox' );
 exports.chrome = exports[ 'google-chrome' ] = require( './chrome' );
 exports.chromium = require( './chromium' );
-exports.canary = exports[ 'chrome-canary' ] = exports[ 'google-chrome-canary' ] = require( './canary' );
+exports['chrome-canary'] = require( './chrome-canary' );
 exports.opera = require( './opera' );
 exports.phantomjs = require( './phantomjs' );
