+++ lib/detect.js
@@ -2,65 +2,50 @@ var spawn = require( 'child_process' ).spawn,
 	winDetect = require( 'win-detect-browsers' ),
 	darwin = require( './darwin' ),
 	extend = require( 'lodash' ).extend,
-	browsers = {
-		'google-chrome': {
-			name: 'chrome',
-			re: /Google Chrome (\S+)/,
-			type: 'chrome',
-			profile: true
+	browsers = { // List of commands and version regex for Linux and Mac
+		chrome: {
+			regex: /Google Chrome (\S+)/,
+			profile: true,
+			variants: {
+				'chrome': [ 'google-chrome', 'google-chrome-stable' ],
+				'chrome-beta': 'google-chrome-beta',
+				'chrome-canary': 'google-chrome-canary'
+			}
 		},
-		'google-chrome-canary': {
-			name: 'canary',
-			re: /Google Chrome (\S+)/,
-			type: 'chrome',
-			profile: true
+		chromium: {
+			regex: /Chromium (\S+)/,
+			profile: true,
+			variants: {
+				'chromium': [ 'chromium', 'chromium-browser' ]
+			}
 		},
-		'chromium': {
-			name: 'chromium',
-			re: /Chromium (\S+)/,
-			type: 'chrome',
-			profile: true
+		firefox: {
+			regex: /Mozilla Firefox (\S+)/,
+			profile: true,
+			variants: {
+				'firefox': 'firefox',
+				'firefox-developer': 'firefox-developer'
+			}
 		},
-		'chromium-browser': {
-			name: 'chromium',
-			re: /Chromium (\S+)/,
-			type: 'chrome',
-			profile: true
-		},
-		'firefox': {
-			name: 'firefox',
-			re: /Mozilla Firefox (\S+)/,
-			type: 'firefox',
-			profile: true
-		},
-		'phantomjs': {
-			name: 'phantomjs',
-			re: /(\S+)/,
-			type: 'phantom',
+		phantomjs: {
+			regex: /(\S+)/,
+			profile: false,
 			headless: true,
-			profile: false
+			variants: 'phantomjs'
 		},
-		'safari': {
-			name: 'safari',
-			type: 'safari',
-			profile: false
+		safari: {
+			profile: false,
+			variants: 'safari'
 		},
-		'ie': {
-			name: 'ie',
-			type: 'ie',
-			profile: false
+		ie: {
+			profile: false,
+			variants: 'ie'
 		},
-		'opera': {
-			name: 'opera',
-			re: /Opera (\S+)/,
-			type: 'opera',
-			image: 'opera.exe',
-			profile: true
+		opera: {
+			regex: /Opera (\S+)/,
+			profile: true,
+			variants: 'opera'
 		}
-	},
-	winDetectMap = {
-		chrome: 'google-chrome',
-		chromium: 'chromium-browser'
 	};
 
 /**
